+++ src/__tests__/directory.ts
@@ -1,5 +1,11 @@
+import { find } from "ramda";
 import { AFile } from "../AFile";
+import { AFileAwait, VirtualFileNotFound } from "../AFileAwait";
 import { AFileSystem } from "../AFileSystem";
+
+import {ReadableStream} from 'web-streams-polyfill/ponyfill/es2018';
+(globalThis as any)['ReadableStream'] = ReadableStream;
+
 describe('AEFunction', () => {
     describe('creation', () => {
         test('CreateEmpty',() =>
@@ -18,33 +24,33 @@ describe('AEFunction', () => {
     describe('find', () => {
         describe('empty', () => {
             const sut = new AFileSystem({}, {readOnly: true, name: 'SUT'});
-            test('empty in root', async () =>
+            test('empty in root', () =>
                 expect(sut.find('/nofile'))
-                    .toBeInstanceOf(Promise));
-            test('empty in root', async () =>
-                expect(await sut.find('/nofile'))
-                    .toBeNull());
+                    .toBeInstanceOf(AFileAwait));
+            test('empty in root', () =>
+                expect(sut.find('/nofile').target)
+                    .resolves.toBeNull());
 
-            test('empty in subdir', async () =>
-                expect(await sut.find('/nodir/nofile'))
-                    .toBeNull());
+            test('empty in subdir', () =>
+                expect(sut.find('/nodir/nofile').target)
+                    .resolves.toBeNull());
 
             // Repeat the tests without a leading slash.
-            test('noslash empty in root', async () =>
-                expect(await sut.find('nofile'))
-                    .toBeNull());
+            test('noslash empty in root', () =>
+                expect(sut.find('nofile').target)
+                    .resolves.toBeNull());
 
-            test('noslash empty in subdir', async () =>
-                expect(await sut.find('nodir/nofile'))
-                    .toBeNull());
+            test('noslash empty in subdir', () =>
+                expect(sut.find('nodir/nofile').target)
+                    .resolves.toBeNull());
 
-            test('latest', async () =>
-                expect(await sut.find('nodir/nofile:latest'))
-                    .toBeNull());
+            test('latest', () =>
+                expect(sut.find('nodir/nofile:latest').target)
+                    .resolves.toBeNull());
 
-            test('earliest', async () =>
-                expect(await sut.find('nodir/nofile:earliest'))
-                    .toBeNull());
+            test('earliest', () =>
+                expect(sut.find('nodir/nofile:earliest').target)
+                    .resolves.toBeNull());
         });
         describe('simple', () => {
             const file1 = new AFile('file1', 'data1');
@@ -59,36 +65,44 @@ describe('AEFunction', () => {
                 }
             }, {readOnly: true, name: 'SUT-Simple'});
             test('inRoot', async () =>
-                expect(await sut.find('/aFile'))
+                expect(await sut.find('/aFile').target)
                     .toBe(file1));
             test('inSubdir', async () =>
-                expect(await sut.find('/subdir/anotherFile'))
+                expect(await sut.find('/subdir/anotherFile').target)
                     .toBe(file2));
             test('Deeper', async () =>
-                expect(await sut.find('/subdir/deeper/versions'))
+                expect(await sut.find('/subdir/deeper/versions').target)
                     .toBe(file2));
             test('Deeper@latest', async () =>
-                expect(await sut.find('/subdir/deeper/versions@latest'))
+                expect(await sut.find('/subdir/deeper/versions@latest').target)
                     .toBe(file2));
             test('Deeper@earliest', async () =>
-                expect(await sut.find('/subdir/deeper/versions@earliest'))
+                expect(await sut.find('/subdir/deeper/versions@earliest').target)
                     .toBe(file1));
             test('Deeper@2', async () =>
-                expect(await sut.find('/subdir/deeper/versions@2'))
+                expect(await sut.find('/subdir/deeper/versions@2').target)
                     .toBe(file2));
             test('Deeper@1', async () =>
-                expect(await sut.find('/subdir/deeper/versions@1'))
+                expect(await sut.find('/subdir/deeper/versions@1').target)
                     .toBe(file1));
 
             test('Deeper@-1', async () =>
-                expect(await sut.find('/subdir/deeper/versions@-1'))
+                expect(await sut.find('/subdir/deeper/versions@-1').target)
                     .toBe(file2));
             test('Deeper@-2', async () =>
-                expect(await sut.find('/subdir/deeper/versions@-2'))
+                expect(await sut.find('/subdir/deeper/versions@-2').target)
                     .toBe(file1));
             test('Deeper@-3', async () =>
-                expect(await sut.find('/subdir/deeper/versions@-3'))
+                expect(await sut.find('/subdir/deeper/versions@-3').target)
                     .toBe(null));
+
+            test('check error', async () =>
+                 expect(sut.find('/noFile').exists.json())
+                    .rejects.toBeInstanceOf(VirtualFileNotFound));
+
+            test('check resolves', async () =>
+                expect(sut.find('/aFile').exists.text())
+                    .resolves.toBe('data1'));
         });
     });
 });